cleanup: move socket name generation into the lib
