+++ src/transport/interprocess.rs
@@ -1,7 +1,7 @@
 //! Custom quinn transport that uses the interprocess crate to provide
 //! local interprocess communication via either Unix domain sockets or
 //! Windows named pipes.
-use std::{io, net::SocketAddr};
+use std::{ffi::OsString, io, net::SocketAddr, path::Path};
 
 use super::quinn_flume_socket::{make_endpoint, FlumeSocket, Packet};
 use bytes::{Buf, Bytes, BytesMut};
@@ -27,6 +27,24 @@ impl<'a> Iterator for FrameIter<'a> {
         }
         self.0.advance(2);
         Some(self.0.split_to(len).freeze())
+    }
+}
+
+/// Automatically chooses name type based on OS support and preference.
+pub fn new_socket_name(root: impl AsRef<Path>, id: &str) -> OsString {
+    let namespaced = {
+        use interprocess::local_socket::NameTypeSupport;
+        let nts = NameTypeSupport::query();
+        match nts {
+            NameTypeSupport::OnlyPaths | NameTypeSupport::Both => false,
+            NameTypeSupport::OnlyNamespaced => true,
+        }
+    };
+
+    if namespaced {
+        format!("@quic-rpc-socket-{}.sock", id).into()
+    } else {
+        root.as_ref().join(format!("{id}.sock")).into()
     }
 }
